Major changes to the arch definition format
Arch definitions automatically construct arbitrary blocks

Blocks are defined by a series of layers whose names correspond to the
appropriate module name (including custom modules defined in
architecture.modules) specified in the definition file. Blocks are
created as nn.Sequential.

Thus, the ConvBlock module is now gone. Attention is now its own module
that has to be added in the definition file.

Parametrizations such as spectral_norm or init methods are
now supported with a special "apply" layer that applies the specified
function to the previous layer in the same block.

TODO : Further support for user defined custom modules